Optimizes the project by deliting unnecessary components
