+++ src/components/app/app.jsx
@@ -1,10 +1,47 @@
-import React from "react";
-import Main from "../main/main.jsx";
+import React, {PureComponent} from "react";
+import Places from "../places/places.jsx";
+import Locations from "../locations/locations.jsx";
 
-const App = () => {
-  return (
-    <Main/>
-  );
-};
+class App extends PureComponent {
+  render() {
+    return (
+      <div className="page page--gray page--main">
+        <header className="header">
+          <div className="container">
+            <div className="header__wrapper">
+              <div className="header__left">
+                <a className="header__logo-link header__logo-link--active">
+                  <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
+                </a>
+              </div>
+              <nav className="header__nav">
+                <ul className="header__nav-list">
+                  <li className="header__nav-item user">
+                    <a className="header__nav-link header__nav-link--profile" href="#">
+                      <div className="header__avatar-wrapper user__avatar-wrapper">
+                      </div>
+                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+                    </a>
+                  </li>
+                </ul>
+              </nav>
+            </div>
+          </div>
+        </header>
+        <main className="page__main page__main--index">
+          <h1 className="visually-hidden">Cities</h1>
+          <div className="tabs">
+            <section className="locations container">
+              <Locations/>
+            </section>
+          </div>
+          <div className="cities">
+            <Places/>
+          </div>
+        </main>
+      </div>
+    );
+  }
+}
 
 export default App;
